fixed the problem with displaying address of house
