+++ src/store/invitations/mutations.ts
@@ -3,7 +3,17 @@ import { InvitationsStateInterface, InvitationsMutationsEnum, Mutations } from '
 
 export const mutations: MutationTree<InvitationsStateInterface> & Mutations = {
   [InvitationsMutationsEnum.CREATE_INVITATION]: (state, payload) => {
-    state.invitations?.push(payload);
+    // const invitation = {
+    //   apartment: {
+    //     ...payload.apartment,
+    //     address: payload.houseAddress,
+    //   },
+    //   email: payload.email,
+    //   id: payload.id,
+    //   status: payload.status,
+    //   type: payload.type,
+    // };
+    // state.invitations?.push(invitation);
   },
   [InvitationsMutationsEnum.SET_APARTMENT_INVITATIONS]: (state, payload) => {
     state.invitations = payload;
